add search by date

+++ src/main/java/com/github/eduumach/hospitalsystem/dto/request/HeartRateRequest.java
@@ -13,7 +13,7 @@ import lombok.experimental.Accessors;
 public class HeartRateRequest {
 
     private String cpf;
-    private int epoch;
+    private Long epoch;
     private double heartRate;
 
     public HeartRateEntity requestObject(){
